requirements.txt
django-filter=23.1
django-cors-headers=3.14.0
Django=4.1.7
python-dotenv=0.21.1
mysqlclient=2.1.1
djangorestframework-simplejwt=5.2.2
djangorestframework=3.14.0
celery=5.2.7
django-redis=5.2.0